Update Query.md
Add in a description of hardcoded variables you can use in queries, seems pretty relevant for bullet journal style scripting of tasks. 

Since this is copied from core/templates, and likely to change(?) maybe it would be best to use an #include directive and break out into a separate file?