only prevent default event on links
I think I misunderstood what was going on originally, mouse and click
weren't both triggering  it was the default event from the a reference.

So, only prevent default behavior on a e;ements, since default handlers
are needed on navigation.

Oddly, this I've only noticed this fixing navigation on firefox simulator.
It wasn't needed on firefox.