Pre-fill new page name with heading
When extracting a page, this will check if the first line is a markdown
heading. If it is, it will be used as a default name in the prompt
asking for a page name.

This uses a regular expression to check for a heading. It seemed a bit
overkill to do a full markdown parsing for this specific check.