Make tag auto-complete context independent
Previously there was a (best effort) attempt to limit tag auto complete to only the type of object a had been applied to (e.g. an item, a page, a task). This is now removed, because I realized the need to reuse tags for various purposes and this filtering was annoying in practice.