Fixes issue when sycing with document editor open
If one had a document editor open and switched to sync mode, silverbullet would reach a deadlock state. This was because the space wasn't fully synced, so the plug for the document editor wasn't available. Silverbullet would respond by redirecting to the raw version of the the document, which would cancel the sync and thus cause an unrecoverable state.