A few fixes for document editors (#1310)
* Fixes a function I forgot in a rework in #1289

* Don't cache the "server-side rendered" pages.

If one would reload in the right moment the browser could respond with the cached version for document editors, which would lead to issues, as it's html.

* Fix spelling: `initalLoad` to `initialLoad`

* Do check that's already done in `updatePageListCache` in `updateDocumentListCache`

* Fixes issue when sycing with document editor open

If one had a document editor open and switched to sync mode, silverbullet would reach a deadlock state. This was because the space wasn't fully synced, so the plug for the document editor wasn't available. Silverbullet would respond by redirecting to the raw version of the the document, which would cancel the sync and thus cause an unrecoverable state.