Fix page decoration prefix appearing in table of content's links
Table of content's links contain page decoration prefixes when navigating between pages or when reloading the system (not on page refresh for some reasons).

This proposal avoids adding those prefixes when the link is having a `pos` in the file which indicates that it targets a heading. I think that it's an expected behavior as the heading could as well be prefixed with something (e.g., an emoji).

Note that it will as well change link's mention widget behavior.