Handle atomic updates for the binary file.
This should fix #1628 as it makes sure to replace the existing `silverbullet`
binary with what gets extracted from the archive instead of just writing to the
binary file directly (which would return a `text file busy` error).

I'm using [`renameio`](https://pkg.go.dev/github.com/google/renameio) to make
sure that the temporary file created during archive extraction is atomically
replacing the target file. In case of an error, the old `silverbullet`
executable should still be there.

The alternative would be to manually create a temporary file and then write to
it when extracting from the archive. After this, we need to atomically rename
this temporary file to the target binary. But, this is effectively what
`renameio` does; the only difference is that `renameio` handles all the corner
cases and possible errors for us.

Another alternative is to unlink (remove) the existing binary first before
extracting the archive. Since the archive extraction creates a new file, this
will be a new inode, while the running binary can still point to the one
started from the original binary. When the update process finishes, the
operating system will collect the reference to the old binary. This has a
problem that we might get an error in between removing the existing binary and
extracting it from the archive, which would result in the upgrade process
leaving us with no `silverbullet` binary to run the update process from. Users
would then need to start a new download from GitHub, etc., which is not
something we'd want.

Furthermore, this last alternative requires us to always keep track of what
files are in the archive and could be loaded during execution of `silverbullet
upgrade`. All of these would need to be unlinked before being extracted from
the archive, otherwise we'd still get the `text file busy` error.

Signed-off-by: Mihai Maruseac <mihai.maruseac@gmail.com>