Make the link to the site point to real site.
It used to point to `v2.silverbullet.md`, but not that all development
is on v2 this is no longer needed? Keeping `v2` means landing on a page
where no dynamic element gets rendered.

In other words, both https://v2.silverbullet.md/ and
https://silverbullet.md/ have similar content as a markdown file, but
only the latter renders the output of queries, templates, etc.

I did an `rg v2` to find other places where this could be changed, but
none seemed relevant.

Signed-off-by: Mihai Maruseac <mihaimaruseac@google.com>